[Add] Credit tab
